// digital-car-museum/autovault-frontend/src/app/shared/models/manufacturer.model.ts
export interface Manufacturer {
  id: number;
  name: string;
  slug: string;
  foundedYear: number;
  founderName: string;
  headquarters: string;
  logoUrl: string;
  history: string;
  description: string;
  websiteUrl: string;
  countryId: number;
  countryName: string;
  countryCode: string;
  carCount: number;
}